Refactor: 바우처 컨트롤러(Runnable) 네이밍 수정
